# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation des APIs publiques.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/cart/add/{id}": {
            "post": {
                "summary": "Ajoute un produit au panier.",
                "operationId": "post_cart_add",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/cart/remove/{id}": {
            "post": {
                "summary": "Supprime un produit du panier.",
                "operationId": "post_cart_remove",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
        "/product/{id}": {
            "get": {
                "summary": "D\u00e9tails d\u2019un produit sp\u00e9cifique.",
                "operationId": "get_product_detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "put": {
                "summary": "D\u00e9tails d\u2019un produit sp\u00e9cifique.",
                "operationId": "put_product_detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "post": {
                "summary": "D\u00e9tails d\u2019un produit sp\u00e9cifique.",
                "operationId": "post_product_detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "summary": "D\u00e9tails d\u2019un produit sp\u00e9cifique.",
                "operationId": "delete_product_detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "options": {
                "summary": "D\u00e9tails d\u2019un produit sp\u00e9cifique.",
                "operationId": "options_product_detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "head": {
                "summary": "D\u00e9tails d\u2019un produit sp\u00e9cifique.",
                "operationId": "head_product_detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "summary": "D\u00e9tails d\u2019un produit sp\u00e9cifique.",
                "operationId": "patch_product_detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "trace": {
                "summary": "D\u00e9tails d\u2019un produit sp\u00e9cifique.",
                "operationId": "trace_product_detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
...
}